Update outdated logic & add few new Tests

+++ src/main/java/krematos/model/ExternalApiRequest.java
@@ -3,11 +3,14 @@ package krematos.model;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
+
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ExternalApiRequest {
     private String transactionId;
-    private double amount;
+    private BigDecimal amount;
     private String currency;
 }
